+++ packages/editor/src/hooks/index.js
@@ -1,5 +1,6 @@
 import './custom-sources-backwards-compatibility';
 import './default-autocompleters';
+import './attach-media-in-post';
 import './media-upload';
 import './pattern-overrides';
 import './navigation-link-view-button';